refactor(cli): move _click into the cli package
Only cli modules imported it, but it sat at the package root. PrintChoice
went to logger.py, whose set_level is its real consumer and which is not a
click construct, so no root module now imports click or the cli package.
