+++ rekordbox_edit/cli/convert.py
@@ -5,8 +5,7 @@ import sys
 
 import click
 
-from rekordbox_edit._click import (
-    PrintChoice,
+from rekordbox_edit.cli._click import (
     add_click_options,
     convert_click_options,
     global_click_confirmations,
@@ -28,7 +27,7 @@ from rekordbox_edit.cli._utils import (
     with_database,
 )
 from rekordbox_edit.display import PrintableField, print_track_info
-from rekordbox_edit.logger import get_debug_file_path, set_level
+from rekordbox_edit.logger import PrintChoice, get_debug_file_path, set_level
 from rekordbox_edit.models import ConvertRequest
 
 logger = logging.getLogger(__name__)
