+++ rekordbox_edit/cli/edit.py
@@ -4,8 +4,7 @@ import logging
 
 import click
 
-from rekordbox_edit._click import (
-    PrintChoice,
+from rekordbox_edit.cli._click import (
     add_click_options,
     edit_click_options,
     global_click_confirmations,
@@ -27,7 +26,7 @@ from rekordbox_edit.cli._utils import (
     with_database,
 )
 from rekordbox_edit.display import PrintableField, print_track_info
-from rekordbox_edit.logger import get_debug_file_path, set_level
+from rekordbox_edit.logger import PrintChoice, get_debug_file_path, set_level
 from rekordbox_edit.models import EditRequest
 
 logger = logging.getLogger(__name__)
