+++ rekordbox_edit/cli/search.py
@@ -4,8 +4,7 @@ import logging
 
 import click
 
-from rekordbox_edit._click import (
-    PrintChoice,
+from rekordbox_edit.cli._click import (
     add_click_options,
     global_click_filters,
     print_option,
@@ -20,7 +19,7 @@ from rekordbox_edit.cli._utils import (
     with_database,
 )
 from rekordbox_edit.display import print_track_info
-from rekordbox_edit.logger import get_debug_file_path, set_level
+from rekordbox_edit.logger import PrintChoice, get_debug_file_path, set_level
 from rekordbox_edit.models import SearchRequest
 
 logger = logging.getLogger(__name__)
